move reviewDetail to separate reducer, fix addon review detail fetching + sherlocked
