+++ src/media/js/app.js
@@ -29,6 +29,7 @@ import WebsiteSubmit from './website/containers/submit';
 
 import addonDashboard from './addon/reducers/dashboard';
 import addonReview from './addon/reducers/review';
+import addonReviewDetail from './addon/reducers/reviewDetail';
 import addonSubmit from './addon/reducers/submit';
 import apiArgs from './site/reducers/apiArgs';
 import login from './site/reducers/login';
@@ -43,6 +44,7 @@ const reducer = combineReducers({
   // The name of the reducers, as imported, will be the keys of state tree.
   addonDashboard,
   addonReview,
+  addonReviewDetail,
   addonSubmit,
   apiArgs,
   login,
